Tested context switching, rtos works, needs refactoring
Everything works, but serious refactoring is needed. Added FPU support.

+++ inc/tick.h
@@ -26,9 +26,9 @@ int gt_get_cntp_ctl();
 int gt_get_cntv_tval();
 int gt_get_cntv_ctl();
 
-int gt_get_cntvct();
+long long gt_get_cntvct();
 int gt_get_cntp_cval();
-int gt_get_cntv_cval();
+long long gt_get_cntv_cval();
 
 int gt_get_cntvoff();
